requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
arxiv==2.2.0
asgiref==3.9.1
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blockbuster==1.5.25
boto3-stubs==1.42.17
botocore-stubs==1.42.17
cachetools==6.1.0
catboost==1.2.8
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
cryptography==44.0.3
cssselect==1.3.0
cssselect2==0.8.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
distro==1.9.0
dj-database-url==3.1.2
Django==5.2.4
executing==2.2.0
faiss-cpu==1.11.0.post1
fastuuid==0.14.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fonttools==4.59.0
forbiddenfruit==0.1.4
freetype-py==2.5.1
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.21
greenlet==3.2.3
groq==0.30.0
grpcio==1.67.1
gunicorn==25.1.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.1
iniconfig==2.3.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jsonschema_rs==0.29.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.79
langchain-groq==0.3.6
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.8
langgraph==0.5.3
langgraph-api==0.2.98
langgraph-checkpoint==2.1.1
langgraph-cli==0.3.5
langgraph-prebuilt==0.5.2
langgraph-runtime-inmem==0.6.0
langgraph-sdk==0.1.73
langsmith==0.4.8
litellm==1.80.11
lxml==6.0.0
lxml_html_clean==0.4.2
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
matplotlib-inline==0.1.7
mcp==1.25.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
mypy-boto3-bedrock-runtime==1.42.3
mypy_extensions==1.1.0
mysqlclient==2.2.8
narwhals==1.47.1
nest-asyncio==1.6.0
networkx==3.5
newspaper3k==0.2.8
ngrok==1.6.0
nltk==3.9.1
numpy==2.3.1
opik==1.9.66
orjson==3.11.0
ormsgpack==1.10.0
oscrypto==1.3.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
playwright==1.55.0
plotly==6.2.0
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==21.0.0
pycairo==1.29.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==6.1.1
pytest==9.0.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.21
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.0
RapidFuzz==3.14.3
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.3
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.2.0
rlPyCairo==0.4.0
rpds-py==0.26.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.1
seaborn==0.13.2
sentence-transformers==5.1.1
sentry-sdk==2.48.0
setuptools==78.1.1
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.47.1
streamlit==1.53.1
structlog==25.4.0
svglib==1.6.0
sympy==1.14.0
tavily-python==0.7.10
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.5.1
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
truststore==0.10.1
types-awscrt==0.30.0
types-s3transfer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.12.0
wikipedia==1.4.0
xhtml2pdf==0.2.17
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
